ArrayUtil2

+++ src/homework/arrayutil/ArrayUtil2.java
@@ -72,7 +72,7 @@ public class ArrayUtil2 {
     }
 
     // eighth
-    int average(int[] array) {
+    double average(int[] array) {
         int average = 0;
         for (int i = 0; i < array.length; i++) {
             average += array[i];
